# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RT Management API",
        "description": "API documentation for RT Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/pengeluaran": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Get list of pengeluaran",
                "operationId": "8617986b048d1402656c508cf15171da",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Create new pengeluaran",
                "operationId": "92df6fff68b8e9855af8d143d16c0a32",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_pengeluaran",
                                    "jumlah",
                                    "bulan",
                                    "tahun",
                                    "tanggal_pengeluaran"
                                ],
                                "properties": {
                                    "nama_pengeluaran": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "bulan": {
                                        "type": "integer"
                                    },
                                    "tahun": {
                                        "type": "integer"
                                    },
                                    "tanggal_pengeluaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/pengeluaran/{id}": {
            "get": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Get detail pengeluaran",
                "operationId": "5d688ba2ce9f2269d97e13b5a91ec38a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Update pengeluaran",
                "operationId": "84b37bf124ac4eaa25bba7a6d35fcf36",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_pengeluaran": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "bulan": {
                                        "type": "integer"
                                    },
                                    "tahun": {
                                        "type": "integer"
                                    },
                                    "tanggal_pengeluaran": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "keterangan": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pengeluaran"
                ],
                "summary": "Delete pengeluaran",
                "operationId": "d9e7fd7dcdb469201561316555edfd7f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
        "/api/penghuni": {
            "get": {
                "tags": [
                    "Penghuni"
                ],
                "summary": "Get list of penghuni",
                "operationId": "22fa370a08a7ac1a893b3002edb511c0",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Penghuni"
                ],
                "summary": "Create new penghuni",
                "operationId": "a6b25842b3df3a54ca5a6bdb9a4f3e84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lengkap",
                                    "status_penghuni",
                                    "nomor_telepon",
                                    "status_menikah"
                                ],
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "foto_ktp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status_penghuni": {
                                        "type": "boolean"
                                    },
                                    "nomor_telepon": {
                                        "type": "string"
                                    },
                                    "status_menikah": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/penghuni/{id}": {
            "get": {
                "tags": [
                    "Penghuni"
                ],
                "summary": "Get detail penghuni",
                "operationId": "b5b93cab33e69a067845caec5dbe8c52",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Penghuni"
                ],
                "summary": "Update penghuni",
                "operationId": "3b0a2a960e79edc4355ce07d223bd3a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "foto_ktp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "status_penghuni": {
                                        "type": "boolean"
                                    },
                                    "nomor_telepon": {
                                        "type": "string"
                                    },
                                    "status_menikah": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Penghuni"
                ],
                "summary": "Delete penghuni",
                "operationId": "818f670d4d7abbef5d5132370495813d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}